% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsoda.R
\docType{package}
\name{lsoda}
\alias{lsoda-package}
\alias{lsoda}
\title{lsoda: C++ Header Library for Ordinary Differential Equations}
\description{
To learn more about `libsoda-cxx` visit the [website](https://github.com/dilawar/libsoda-cxx).
}
\keyword{internal}
