% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ode}
\alias{ode}
\title{Ordinary differential equation solver using lsoda}
\usage{
ode(y, times, func, rtol = 1e-06, atol = 1e-06)
}
\arguments{
\item{y}{vector of initial state values}

\item{times}{vector of times -- including the start time}

\item{rtol}{double for the relative tolerance}

\item{atol}{double for the absolute tolerance}

\item{f}{R function with signature function(t,y)}
}
\value{
a matrix for times in the first column and the state values in the other columns.
}
\description{
Ordinary differential equation solver using lsoda
}
